+++ app/src/main/java/com/example/asssignmentsdktesttask/data/network/deserializers/LatestRateResponseDeserializer.kt
@@ -18,7 +18,7 @@ class LatestRateResponseDeserializer @Inject constructor() :
     ): LatestRateResponse {
         val jsonObject = json.asJsonObject
 
-        val timestamp = jsonObject.get("timestamp").asInt
+        val timestamp = jsonObject.get("timestamp").asLong
         val date = jsonObject.get("date").asString
 
         if (jsonObject.get("success").asBoolean) {
